requirements.txt
Django==1.7
Pillow==2.5.3
argparse==1.2.1
dj-database-url==0.3.0
django-cape==0.2
psycopg2==2.5.4
wsgiref==0.1.2
django-exceptional-middleware==1.2.1
factory-boy==2.4.1
django-webtest==1.7.7
webtest==2.0.16
django-database-constraints==0.2
